--- src/components/EyesOff.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import eyesOffImg from '../assets/noeyes.svg';
-
-function EyesOff() {
-    return (
-        <>
-            <img src={eyesOffImg} alt="비밀번호 보기" />
-        </>
-    );
-}
-
-export default EyesOff;
