--- src/components/EyesOn.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import eyesOnImg from '../assets/eyes.svg';
-
-function EyesOn() {
-    return (
-        <>
-            <img src={eyesOnImg} alt="비밀번호 보기" />
-        </>
-    );
-}
-
-export default EyesOn;
