+++ src/components/Gnb.js
@@ -1,10 +1,11 @@
 import React, { useEffect, useState } from 'react';
-import Logo from './Logo';
-import CTA from './Cta';
+import Logo from './logo';
+import CTA from './cta';
 import { Link } from 'react-router-dom';
 import profileImg from '../assets/profile.svg';
-import { getProfile } from '../api/profile';
+import { getProfile } from '../api/api';
 import '../styles/gnb.css';
+
 function GNB() {
     const [data, setData] = useState();
 
@@ -26,25 +27,23 @@ function GNB() {
     }, []);
 
     return (
-        <>
-            <header>
-                <div className="header-container">
-                    <Link to="/">
-                        <Logo />
+        <header>
+            <div className="header-container">
+                <Link to="/">
+                    <Logo />
+                </Link>
+                {data ? (
+                    <div className="header-profile">
+                        <img src={profileImg} alt="프로필" />
+                        <p>{data.email}</p>
+                    </div>
+                ) : (
+                    <Link to="/sign-in" style={LinkStyle}>
+                        <CTA text={'로그인'} />
                     </Link>
-                    {!data ? (
-                        <Link to="/sign-in" style={LinkStyle}>
-                            <CTA text={'로그인'} />
-                        </Link>
-                    ) : (
-                        <div className="header-profile">
-                            <img src={profileImg} alt="프로필" />
-                            <p>{data.email}</p>
-                        </div>
-                    )}
-                </div>
-            </header>
-        </>
+                )}
+            </div>
+        </header>
     );
 }
