+++ src/components/Star.js
@@ -3,21 +3,21 @@ import StarDefault from '../assets/starDefault.svg';
 import StarSeleted from '../assets/starSeleted.svg';
 
 function Star() {
-    const [starState, setStarState] = useState('default');
+    const [isStarShowing, setIsStarShowing] = useState(false);
 
-    const handleStarSeleted = (e) => {
-        if (starState === 'default') {
-            setStarState('seleted');
+    const handleStarSeleted = () => {
+        if (isStarShowing === false) {
+            setIsStarShowing(true);
         } else {
-            setStarState('default');
+            setIsStarShowing(false);
         }
     };
     return (
         <>
-            {starState === 'default' ? (
-                <img onClick={handleStarSeleted} src={StarDefault} alt="즐겨찾기" />
-            ) : (
+            {isStarShowing ? (
                 <img onClick={handleStarSeleted} src={StarSeleted} alt="즐겨찾기 해제" />
+            ) : (
+                <img onClick={handleStarSeleted} src={StarDefault} alt="즐겨찾기" />
             )}
         </>
     );
