+++ src/styles/Landing.css
@@ -62,7 +62,7 @@
     background: linear-gradient(180deg, rgba(196, 196, 196, 0) 67.68%, #f0f6ff 94.76%);
 }
 
-.a {
+.interval {
     padding-top: 120px !important;
 }
