+++ src/styles/shard.css
@@ -37,7 +37,7 @@
     line-height: normal;
 }
 .shared-content {
-    padding: 40px 190px;
+    padding: 40px 350px;
     display: flex;
     flex-direction: column;
     gap: 20px;
